+++ src/f3dasm/simulation/__init__.py
@@ -1,3 +1,3 @@
-from .abaqus import *
+from .abaqus_simulator import AbaqusSimulator
+from .abaqus_utils import *
 from .simulator import Simulator
-
