--- src/f3dasm/simulation/abaqus/__init__.py
@@ -1,2 +1,0 @@
-from .abaqus_simulator import *
-from .abaqus_utils import *
